requirements.txt
flask
flask-cors
python-dotenv
requests
beautifulsoup4
openai
httpx